+++ src/BugsnagUnity/IConfiguration.cs
@@ -13,6 +13,8 @@ namespace BugsnagUnity
 
         string[] DiscardClasses { get; set; }
 
+        string[] ProjectPackages { get; set; }
+
         bool ErrorClassIsDiscarded(string className);
 
         ErrorTypes[] EnabledErrorTypes { get; set; }
@@ -69,5 +71,7 @@ namespace BugsnagUnity
         ulong AppHangThresholdMillis { get; set; }
 
         string BundleVersion { get; set; }
+
+        string AppType { get; set; }
     }
 }